pyproject.toml
[tool.black]
line-length = 119
target-version = ["py38"]

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true

line_length = 119
lines_after_imports = 2
multi_line_output = 3
use_parentheses = true

[tool.poetry]
name = "chess-bot"
version = "0.1.0"
description = ""
authors = ["Jeongho Ju <jeongho@comcom.ai>", "YoungJae Kim <byeongal@comcom.ai>", "Donghoon Baek <baek@comcom.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.0.0"
Pillow = "9.2.0"
torch = "1.12.0"
transformers = "4.20.1"
chess = "1.9.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"